+++ protected/humhub/modules/file/controllers/FileController.php
@@ -2,7 +2,7 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2015 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) 2017 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
@@ -33,16 +33,6 @@ class FileController extends \humhub\components\Controller
                 'class' => AccessControl::className(),
                 'guestAllowedActions' => ['download']
             ],
-                /*
-                  'httpCache' => [
-                  'class' => 'yii\filters\HttpCache',
-                  'only' => ['download'],
-                  'cacheControlHeader' => 'public, max-age=31536000',
-                  'etagSeed' => function ($action, $params) {
-                  return serialize([\yii\helpers\Url::current()]);
-                  },
-                  ],
-                 */
         ];
     }