+++ protected/humhub/docs/CHANGELOG.md
@@ -94,6 +94,7 @@ HumHub Change Log
 - Enh: Added option to show/hide deactivated user content in stream
 - Enh: Allow any url route as homepage by homeUrl array application parameter
 - Fix #2255: Added missing Social Account Settings menu
+- Fix: Added missing file download http caching
 
 
 1.2.0-beta.2 (February 24, 2017)